apt: separate migrate_apt_sources from configure_default_apt_sources

+++ uaclient/apt.py
@@ -1,7 +1,6 @@
 import glob
 import logging
 import os
-import re
 import shutil
 
 from uaclient import util
@@ -176,52 +175,65 @@ def remove_apt_list_files(repo_url, series):
             os.unlink(path)
 
 
-def reconfigure_apt_sources(cfg=None, platform_info=None):
-    """Reconfigure apt sources for the given series.
+def migrate_apt_sources(cfg=None, platform_info=None):
+    """Migrate apt sources list files across upgrade/downgrade boundary.
 
-    Automatically setup an unauthenticated apt source for esm
-    production if not already present.
-    For each configured and enabled entitlement, update the sources files to
-    the new series.
+    Only migrate apt sources if we are attached and an entitlement is
+    active. (Meaning they have existing apt policy reference).
 
-    @param platform_info: dict of platform information for testing
+    @param cfg: UAClient configuration instance for testing
+    @param cfg: platform information dict for testing
     """
+
     from uaclient import config
     from uaclient import entitlements
+    from uaclient import status
 
     if not platform_info:  # for testing
         platform_info = util.get_platform_info()
     if not cfg:  # for testing
         cfg = config.UAConfig()
-
-    policy, _err = util.subp(['apt-cache', 'policy'])
+    if not cfg.is_attached:
+        return
     for ent_cls in entitlements.ENTITLEMENT_CLASSES:
         if not hasattr(ent_cls, 'repo_url'):
             continue
-        repo_filename = ent_cls.repo_list_file_tmpl.format(
-            name=ent_cls.name, series=platform_info['series'])
-        repo_list_glob = ent_cls.repo_list_file_tmpl.format(
-            name=ent_cls.name, series='*')
-        repo_url = ent_cls.repo_url
-        if cfg.is_attached:
-            resource_cfg = cfg.entitlements.get(ent_cls.name, {})
-            directives = resource_cfg.get(
-                'entitlement', {}).get('directives', {})
-            repo_url = directives.get('aptURL')
-        enabled = bool(re.search(r'(?P<pin>(-)?\d+) %s' % repo_url, policy))
-        if not os.path.exists(repo_filename):
-            if ent_cls.name == 'esm':
-                if platform_info['series'] == 'trusty':
-                    logging.info(
-                        'Providing unauthenticated ESM apt source file: %s',
-                        repo_filename)
-                    add_auth_apt_repo(repo_filename, repo_url)
-            elif enabled:  # And not esm
-                logging.info('Upgrading existing "%s" apt source files: %s',
-                             ent_cls.title, repo_filename)
-
-                add_auth_apt_repo(repo_filename, repo_url)
+        entitlement = ent_cls(cfg)
+        op_status, _details = entitlement.operational_status()
+        if op_status != status.ACTIVE:
+            continue
+        repo_list_glob = entitlement.repo_list_file_tmpl.format(
+            name=entitlement.name, series='*')
+        # Remove invalid series list files
         for path in glob.glob(repo_list_glob):
             if platform_info['series'] not in path:
                 logging.info('Removing old apt source file: %s', path)
                 os.unlink(path)
+        pass_affordances, details = entitlement.check_affordances()
+        if not pass_affordances:
+            logging.info(
+                'Disabled %s after package upgrade/downgrade. %s',
+                entitlement.title, details)
+        entitlement.enable()  # Re-enable on current series
+
+
+def configure_default_apt_sources(platform_info=None):
+    """Configure any default apt sources for uaclient entitlenents.
+
+    Currently only setup unauthenticated esm on trusty.
+
+    @param platform_info: dict of platform information for testing
+    """
+    from uaclient import entitlements
+
+    if not platform_info:  # for testing
+        platform_info = util.get_platform_info()
+
+    esm_cls = entitlements.ENTITLEMENT_CLASS_BY_NAME['esm']
+    repo_filename = esm_cls.repo_list_file_tmpl.format(
+        name=esm_cls.name, series=platform_info['series'])
+    if platform_info['series'] == 'trusty':
+        if not os.path.exists(repo_filename):
+            logging.info('Providing unauthenticated ESM apt source file: %s',
+                         repo_filename)
+            add_auth_apt_repo(repo_filename, esm_cls.repo_url)
